requirements.txt
alembic==1.13.1
annotated-types==0.5.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
Authlib==1.3.0
Babel==2.12.1
backcall==0.2.0
backports.tarfile==1.0.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
bidict==0.23.1
black==22.6.0
bleach==6.0.0
blinker==1.7.0
blis==0.7.10
boto3==1.34.29
botocore==1.34.29
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
comm==0.1.4
confection==0.1.1
cryptography==42.0.5
cymem==2.0.7
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
dnspython==2.6.1
easy-colors==0.0.1
ecdsa==0.18.0
email_validator==2.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.110.0
FastAPI-SQLAlchemy==0.2.1
fastjsonschema==2.18.0
filelock==3.12.4
Flask==3.0.3
flask-babel==4.0.0
Flask-BabelEx==0.9.4
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fqdn==1.5.1
functorflow==2.2.5
greenlet==3.0.2
h11==0.14.0
idna==3.4
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.25.1
ipython==8.14.0
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
keyring==25.1.0
langcodes==3.3.0
ldap3==2.9.1
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
more-itertools==10.2.0
murmurhash==1.0.9
mypy-extensions==0.4.3
nbclient==0.8.0
nbconvert==7.7.4
nbformat==5.9.2
nest-asyncio==1.5.7
nltk==3.8.1
notebook==7.0.2
notebook_shim==0.2.3
numpy==1.25.2
overrides==7.4.0
packaging==23.1
pandocfilters==1.5.0
paramiko==3.4.0
parso==0.8.3
passlib==1.7.4
pathspec==0.9.0
pathy==0.10.2
pickleshare==0.7.5
platformdirs==3.11.0
preshed==3.0.8
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
psycopg2==2.9.9
pure-eval==0.2.2
pyasn1==0.6.0
pycparser==2.21
pydantic==2.3.0
pydantic-settings==2.2.1
pydantic_core==2.6.3
Pygments==2.16.1
PyNaCl==1.5.0
pyotp==2.9.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.9
python-socketio==5.11.2
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qrcode==7.4.2
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.9.2
rsa==4.9
s3transfer==0.10.0
scikit-learn==1.4.0
scipy==1.11.4
send-mail==0.0.2
Send2Trash==1.8.2
simple-websocket==1.0.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
speaklater==1.3
SQLAlchemy==2.0.23
sqlparse==0.4.4
srsly==2.4.7
sshtunnel==0.4.0
stack-data==0.6.2
starlette==0.36.3
terminado==0.17.1
thinc==8.1.12
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typer==0.9.0
typing_extensions==4.10.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.4
uvicorn==0.29.0
virtualenv==20.24.5
wasabi==1.1.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==3.0.2
wsproto==1.2.0
WTForms==3.1.2
zipp==3.18.1
